Added check for Spoiler level
Tags and traits binding now checks against the Global MaxSpoiler for loading tags/traits.
I changed Spoiler in hopefully all classes to a Byte(since it should only be 0-3)

+++ VisualNovelManagerv2/Design/Settings/UserSettings.cs
@@ -9,7 +9,7 @@ namespace VisualNovelManagerv2.Design.Settings
     public class UserSettings
     {
         public bool NsfwEnabled { get; set; }
-        public uint MaxSpoilerLevel { get; set; }
+        public byte MaxSpoilerLevel { get; set; }
         public VnSetting VnSetting { get; set; }
     }
