+++ VisualNovelManagerv2/Globals.cs
@@ -16,6 +16,6 @@ namespace VisualNovelManagerv2
         public static uint VnId;
         public static readonly StatusBarViewModel StatusBar = (new ViewModelLocator()).StatusBar;
         public static bool NsfwEnabled = false;
-        public static uint MaxSpoiler = 0;
+        public static byte MaxSpoiler = 0;
     }
 }
